server: Extend certificate enrollment status operation to use a wildcard
If the STAUT MAC address is not known (e.g., when MAC address
randomization is used), it can be difficult to get the database cleared
and status fetched during a test sequence. As a workaround, allow a
special ClientMACAddr value "any" to be used to match any entry. This
will not work if there are multiple different test sequences in
parallel, but should work with the specific case of a setup where only a
single test instance can occur at the time.

For example, this type of a sequence could be used to perform a test:

server_reset_default,Program,HS2-R2,ClientMACAddr,any
sta_osu,interface,wlan0
server_request_status,Program,HS2-R2,Device,OSUServer,ClientMACAddr,any,Status,OSU,Timeout,60

Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>