DPP: Check mDNS discovery result on DPPmDNSEnable,Yes
Verify that mDNS discovery of a Controller for AP/Relay has succeeded
when the SetParameter command is used to explicitly enable mDNS
discovery. While the actual discovery operation has already happened
earlier while processing the ap_preset_testparameters command with
DPPConfiguratorAddress,mDNS, there is no convenient mechanism to report
the error for that since it is possible for the TCP/IP stack not being
ready before the ap_config_commit command.

Try mDNS discovery again here if the previous attempt failed and report
a failure if the Controller cannot be discovered.

Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>