traffic: Enable broadcast ping option only for IPv4 pings
The broadcast ping option is not applicable for ping6 (at least on
Android) even though the usage() has listed it as supported. Thus,
enabling -b option for ping6 may make the command fail. Fix this by
adding the -b argument only for the IPv4 case which actually needs this
for the broadcast ping.

Fixes: 25f27205b792 ("traffic: Allow pinging a broadcast address")
Signed-off-by: Veerendranath Jakkam <quic_vjakkam@quicinc.com>