DPP: More options for discovering the Controller for a Relay
Try mDNS again when ap_config_commit is issued if the Controller address
was not discovered at the time ap_preset_testparameters
DPPConfiguratorAddress,mDNS was issued. In addition, allow hostapd to be
started even if mDNS discovery has not completed. In that case, hostapd
will only accept connections initiated by the Controller.

Allow the ap_preset_testparameters command with
DPPConfiguratorAddress,mDNS to be used after hostapd has been started to
complete Controller discovery. That command still has to be issued
before ap_config_commit, but the second instance of it can perform mDNS
discovery and Controller addition after that.

Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>