Do not add vht_capab for the 2.4 GHz interface in dual band AP
When trying to enable dual AP (2.4 GHz + 5 GHz), 5 GHz vht_capab ended
up overriding 2.4 GHz's which resulted in failure to start the AP
due to 2 GHz VHT capability mismatch/not support.

VHT is not used on the 2.4 GHz band, so remove vht_capab field when
configuring the 2.4 GHz interface parameters.

Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>