DPP: Start unconfigured AP on channel 6 if PB is used
The new rules for DP discovery channels require the AP to be operating
on a preferred chirping channel if Configurator connectivity is not
advertised so that the PB Enrollee can discover it. sigma_dut used
channel 11 by default, if not overridden with an explicit configuration.
Change that to be channel 6 if PB is used.

Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>