traffic: Include -b argument for ping only for 255.255.255.255
By default enabling broadcast pinging causes issues with tests when the
framesize is more than 1472 due to max MTU size limit. This happens with
the -b argument even if the actual IPv4 address is a unicast address. To
avoid such issues enable broadcast pinging with -b only when the IPv4
address is 255.255.255.255.

Fixes: 25f27205b792 ("traffic: Allow pinging a broadcast address")
Signed-off-by: Shivani Baranwal <quic_shivbara@quicinc.com>