NANR4: Configure Unsynchronised Service Discovery parameters
In USD operations of NAN, the device will stay on default publish
channel and advertise FSD configuration parameters in SDEA attribute of
Publish frames in solicited and unsolicited manner as specified by the
user. When the DiscoveryResult event is declared the subscriber will
send a follow up message with or without service specific information.

Configure Publish method and Follow-up method with given FSD parameters
and Service Specific Info.

Signed-off-by: Vinay Gannevaram <quic_vganneva@quicinc.com>