Modify bssid size in strlcpy() to accommodate NULL
Modify the bssid length to be copied in strlcpy(). strlcpy() copies at
most n-1 bytes into the destination buffer, where n is the size of bytes
specified in the third paramter of strlcpy(). Also, it adds a null
terminator following the bytes copied to the destination. Use sizeof()
to get the real length of the buffer instead of hardcoded value.

Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>