iperf: Support latency configuration and reporting with iperf V2
For latency configuration, in server mode, --histogram is used
with LowerCI (Confidence Interval) and UpperCI configuration.
Default LowerCI is 95% and UpperCI is 99.9%.

In client mode, --trip-times is used to enable the measurement of
end to end write to read latencies.

Client and server clocks must be synchronized to a common reference,
e.g., using Network Time Protocol (NTP), to use the latency
configuration.

Configure --realtime option to request real-time scheduler and use
-w 2M to increase the buffer size to 2 MB.

Also, print the iperf command for debug help.

Signed-off-by: Purushottam Kushwaha <quic_pkushwah@quicinc.com>