STA: Add support for dev_upload_log command
Add support for dev_upload_log functionality to stop logging and to
upload the device logs using FTP using the FTP parameters given in the
command.

The dev_upload_log command is added in the EHT program. There are no
existing legacy programs using this dev_upload_log CAPI command. The
older CAPI commands dev_start_test and dev_stop_test that were
introduced to add auto log capture support are not used in any existing
programs and are removed from CAPI specification document.

Restarting the cnss_diag_lite is not needed as it can capture the logs
into the files. If the file is removed, it will create the new file and
captures logs in the newly created file.

Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>