set_wpa_common: Make PMF as optional explicitly for WPA3 program
Currently, when the "AKMSuiteType" parameter is not specified in the
sta_set_security CAPI command, STAUT is expected to configure all
supported AKMs in the network profile. However, STAUT is not
enabling WPA-EAP-SHA256 AKM though it is supported due to PMF is not
configured as optional in the network profile. This is causing
connection issues with AP if it is advertising only WPA-EAP-SHA256 AKM
as known AKM.

To address this issue, make PMF as optional explicitly for the WPA3
program when configuring network profile.

Signed-off-by: Kavita Kavita <quic_kkavita@quicinc.com>