AP: Configure unsolicited broadcast Probe Response interval for 6 GHz
Currently unsolicited broadcast Probe Response interval is not
configured in the 6 GHz band and hence, AP is not sending unsolicited
Probe Response within an interval of 20 TU + 3 TUs.

Configure the unsolicited broadcast probe response interval as 20 TU for
the 6 GHz band and 11ax mode. For non-zero interval, the AP schedules
unsolicited broadcast Probe Response frames to be transmitted for
in-band discovery.

Signed-off-by: Yu Zhang(Yuriy) <yu.zhang@oss.qualcomm.com>